Adjustments to follow guidlines
Added in metadata for rendering modes. Also there's some slight adjustments to type checking.
Seed is now a "private" member so the seed function just sets that, even though it should be deprecated. Also added in the reward range change. Reset now takes in seed and return_info according to the base class. Step had a small syntax change since I realized having multiple if statements made less sense (for now). Type checking for render function as well as handling rendering as an image. Current the trash is white and the garbage is black, and the player is a red pixel. Updates to seed so it properly seeds, however warn that it is deprecated. Also overload the close function so it exists but we don't have anything else to do currently.